Aula 7: Queries
in  the class we've learned how to use parameters in the queries we do to the api:

Static way through function annotation
 @Query("endpoint?param=value&otherparam=othervalue")
No way to change these at runtime. This approach is useful when you need to pass the same params to every single call.

Dynamic way through function's parameter annotation.
   @Get("endpoint")
     fun callToApi(
         @Query("param1") param1: String,
         @Query("param2") param2: Int?
     ): Call<ApiCallResponse>

The second param is an optional parameter,  if you pass null to an optional parameter retrofit won't
put it in the final URl.

The Multi parameter approach
@GET("endpoint")
fun callGetWithMapOfParams(@QueryMap params: Map<String, String>): Call<ApiCallResponse>

Here we can pass a  Map<paramName, paramValue> and retrofif with build the url based on it